+++ src/main/java/com/auction/server/factory/UserFactory.java
@@ -2,6 +2,8 @@ package com.auction.server.factory;
 
 import com.auction.models.*;
 
+import java.math.BigDecimal;
+
 public class UserFactory {
     public static User createNewUser(UserRole role, String username, String password) throws IllegalArgumentException {
         switch (role) {
@@ -18,7 +20,7 @@ public class UserFactory {
         }
     }
 
-    public static User createUserFromDB(String id, UserRole role, String username, String password, double balance)
+    public static User createUserFromDB(String id, UserRole role, String username, String password, BigDecimal balance)
             throws IllegalArgumentException {
         switch (role) {
             case BIDDER -> {
